+++ CHANGELOG.md
@@ -6,10 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
 - Version 2 of device endpoints to fetch list of devices or device by uid
-
+- Device peer recovery `list` method to fetch list of peer recoveries
+- Device peer recovery `set` method to set peer recovery
 
 ## [1.11.0] - 2022-06-03